hostname module already sets /etc/hostname
The hostname module already sets /etc/hostname.
Or the file in /etc/sysconfig/whatever for old RedHat.
And if it doesn't know how to handle the distibution,
the hostname module will fail anyway.

Caveat: this means we no longer enforce correct permissions on
/etc/hostname.  This is significant if the user already configured a
restrictive umask, because the hostname module doesn't account for that.
Document it.